+++ autd3-driver/src/datagram/fpga_gpio_out.rs
@@ -1,7 +1,7 @@
 use std::convert::Infallible;
 
 use crate::firmware::{
-    fpga::{DebugType, GPIOOut},
+    fpga::{GPIOOut, GPIOOutputType},
     operation::DebugSettingOp,
 };
 
@@ -14,21 +14,21 @@ use derive_more::Debug;
 ///
 /// ```
 /// # use autd3_driver::datagram::GPIOOutputs;
-/// # use autd3_driver::firmware::fpga::{DebugType, GPIOOut};
+/// # use autd3_driver::firmware::fpga::{GPIOOutputType, GPIOOut};
 /// GPIOOutputs::new(|dev, gpio| match gpio {
-///     GPIOOut::O0 => DebugType::BaseSignal,
-///     GPIOOut::O1 => DebugType::Sync,
-///     GPIOOut::O2 => DebugType::PwmOut(&dev[0]),
-///     GPIOOut::O3 => DebugType::Direct(true),
+///     GPIOOut::O0 => GPIOOutputType::BaseSignal,
+///     GPIOOut::O1 => GPIOOutputType::Sync,
+///     GPIOOut::O2 => GPIOOutputType::PwmOut(&dev[0]),
+///     GPIOOut::O3 => GPIOOutputType::Direct(true),
 /// });
 /// ```
 #[derive(Debug)]
-pub struct GPIOOutputs<F: Fn(&Device, GPIOOut) -> DebugType + Send + Sync> {
+pub struct GPIOOutputs<F: Fn(&Device, GPIOOut) -> GPIOOutputType + Send + Sync> {
     #[debug(ignore)]
     f: F,
 }
 
-impl<F: Fn(&Device, GPIOOut) -> DebugType + Send + Sync> GPIOOutputs<F> {
+impl<F: Fn(&Device, GPIOOut) -> GPIOOutputType + Send + Sync> GPIOOutputs<F> {
     /// Creates a new [`GPIOOutputs`].
     #[must_use]
     pub const fn new(f: F) -> Self {
@@ -36,11 +36,11 @@ impl<F: Fn(&Device, GPIOOut) -> DebugType + Send + Sync> GPIOOutputs<F> {
     }
 }
 
-pub struct DebugSettingOpGenerator<F: Fn(&Device, GPIOOut) -> DebugType + Send + Sync> {
+pub struct DebugSettingOpGenerator<F: Fn(&Device, GPIOOut) -> GPIOOutputType + Send + Sync> {
     f: F,
 }
 
-impl<F: Fn(&Device, GPIOOut) -> DebugType + Send + Sync> OperationGenerator
+impl<F: Fn(&Device, GPIOOut) -> GPIOOutputType + Send + Sync> OperationGenerator
     for DebugSettingOpGenerator<F>
 {
     type O1 = DebugSettingOp;
@@ -57,7 +57,7 @@ impl<F: Fn(&Device, GPIOOut) -> DebugType + Send + Sync> OperationGenerator
     }
 }
 
-impl<F: Fn(&Device, GPIOOut) -> DebugType + Send + Sync> Datagram for GPIOOutputs<F> {
+impl<F: Fn(&Device, GPIOOut) -> GPIOOutputType + Send + Sync> Datagram for GPIOOutputs<F> {
     type G = DebugSettingOpGenerator<F>;
     type Error = Infallible;